Moved the Cryptocurrency and Cryptography folders to pages directory.
