+++ src/components/pages/cryptocurrency/Calculator.vue
@@ -233,7 +233,7 @@
 
 <script setup lang="ts">
 import { ref, reactive, watch } from 'vue'
-import MarqueeTicker from '../partials/MarqueeTicker.vue'
+import MarqueeTicker from '../../partials/MarqueeTicker.vue'
 
 // Define component name
 defineOptions({
