+++ src/components/pages/cryptocurrency/DomainResolver.vue
@@ -149,9 +149,9 @@
 <script setup lang="ts">
 import { ref, reactive, computed, onMounted } from 'vue'
 import { initFlowbite } from 'flowbite'
-import { domainMixins } from '../../utils/mixins/domainMixins'
+import { domainMixins } from '../../../utils/mixins/domainMixins'
 import { ens } from '@background/functions/utils/domains/ens'
-import DomainResolutionBadge from '../domains/ResolutionBadge.vue'
+import DomainResolutionBadge from '../../domains/ResolutionBadge.vue'
 
 // Define component name
 defineOptions({
