+++ src/components/pages/cryptography/EncryptDecryptTool.vue
