+++ src/components/pages/cryptography/HashTool.vue
